Fixes re bridges
Fixed handling of bridges. It now checks both of the input "tables" so that neither one may be a bridge. Also, the as-is copying of bridge segments now properly uses the split_geom field so the geometry isn't accidentally blanked after the geometry column is renamed.

Also, some readability improvements.